refactor: split tree tests into dedicated modules
